Move win32/win32_compat.h to include/win32/win32_compat.h
