+++ include/win32/win32_compat.h
