+++ lib/libnfs-sync.c
@@ -27,7 +27,7 @@
 #endif
 
 #ifdef WIN32
-#include "win32_compat.h"
+#include <win32/win32_compat.h>
 #endif
 
 #ifdef HAVE_SYS_SOCKET_H
