+++ lib/libnfs-zdr.c
@@ -26,7 +26,7 @@
 #endif
 
 #ifdef WIN32
-#include "win32_compat.h"
+#include <win32/win32_compat.h>
 #endif
 
 #ifdef AROS
