+++ lib/socket.c
@@ -24,7 +24,7 @@
 #endif
 
 #ifdef WIN32
-#include "win32_compat.h"
+#include <win32/win32_compat.h>
 #endif
 
 #ifdef HAVE_ARPA_INET_H
