+++ nfs4/nfs4.c
@@ -17,7 +17,7 @@
 */
 
 #ifdef WIN32
-#include "win32_compat.h"
+#include <win32/win32_compat.h>
 #else
 #include <sys/stat.h>
 #endif/*WIN32*/
