+++ nsm/nsm.c
@@ -16,7 +16,7 @@
 */
 
 #ifdef WIN32
-#include "win32_compat.h"
+#include <win32/win32_compat.h>
 #endif/*WIN32*/
 
 #include <stdio.h>
